requirements.txt
chardet==5.2.0
docstring_parser==0.16
mypy-extensions==1.0.0
packaging==24.2
pillow==11.0.0
pylabels==1.2.1
reportlab==4.2.5
typed-argument-parser==1.10.1
typing-inspect==0.9.0
typing_extensions==4.12.2
